refactor: Move BasicPathTestCase to correct package
