+++ application/src/test/java/org/opentripplanner/routing/algorithm/transferoptimization/api/OptimizedPathTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.raptor.api.model.RaptorValueFormatter;
 import org.opentripplanner.raptorlegacy._data.RaptorTestConstants;
-import org.opentripplanner.raptorlegacy._data.stoparrival.BasicPathTestCase;
+import org.opentripplanner.raptorlegacy._data.api.BasicPathTestCase;
 
 class OptimizedPathTest implements RaptorTestConstants {
