+++ application/src/test/java/org/opentripplanner/routing/algorithm/transferoptimization/model/MinSafeTransferTimeCalculatorTest.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.routing.algorithm.transferoptimization.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.opentripplanner.raptorlegacy._data.stoparrival.BasicPathTestCase.C1_CALCULATOR;
+import static org.opentripplanner.raptorlegacy._data.api.BasicPathTestCase.C1_CALCULATOR;
 import static org.opentripplanner.routing.algorithm.transferoptimization.model.MinSafeTransferTimeCalculator.bound;
 import static org.opentripplanner.utils.time.TimeUtils.time;
 
@@ -9,8 +9,8 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.raptor.api.path.RaptorPath;
 import org.opentripplanner.raptorlegacy._data.RaptorTestConstants;
+import org.opentripplanner.raptorlegacy._data.api.BasicPathTestCase;
 import org.opentripplanner.raptorlegacy._data.api.TestPathBuilder;
-import org.opentripplanner.raptorlegacy._data.stoparrival.BasicPathTestCase;
 import org.opentripplanner.raptorlegacy._data.transit.TestTripSchedule;
 import org.opentripplanner.utils.time.DurationUtils;
